# [PROPOSAL] Introduce SQLite to handle counters